requirements.txt
customtkinter
minecraft-launcher-lib
requests
Pillow
tkvideo
darkdetect
pypresence
